requirements.txt
pylint
pytest
pytest-cov
flake8
black
pandas
requests
azureml-core
